requirements.txt
aiofiles==23.2.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
audiocraft==1.3.0
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
crema==0.2.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
essentia==2.1b6.dev1110
fastapi==0.115.6
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.7
keras==2.8.0
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.11.9
llvmlite==0.43.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mir-eval==0.7
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
nnAudio==0.3.3
numba==0.60.0
numpy==1.26.3
openai-whisper==20240930
optuna==4.1.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
protobuf==3.20.3
psutil==6.1.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.8.4
scikit-learn==1.6.0
scipy==1.13.1
semantic-version==2.10.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tensorflow==2.8.0
tiktoken==0.8.0
tomli==2.2.1
tomlkit==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.6.1
torchvision==0.16.0
tqdm==4.67.1
typeguard==2.13.3
typer==0.15.1
typing_extensions==4.12.2
tables==3.10.1
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.1
websockets==12.0
zipp==3.21.0
pyworld==0.3.5
demucs==4.0.1